Add support for IBM Cloud Orchestrator integration to the Keystone service.
This change adds a relation interface, context generator, templates,
and apppropriate hooks to relate the keystone-ico subordinate charm
(cs:~mskalka/keystone-ico) to the Keystone principal charm.

A new template for /etc/keystone/keystone-paste.ini had to be
created which overrides some defaults from the upstream file. If
the keystone-ico charm is related it will manage these overrides,
otherwise fall back to the default.

Change-Id: Ibf1879d83d6f54e3331bf4de61d8116d1eeff40c